+++ src/control-plane/backend/lambda/api/store/dynamodb/dynamodb-metadata-store.ts
@@ -15,7 +15,6 @@ import {
   UpdateCommand,
   QueryCommandInput,
 } from '@aws-sdk/lib-dynamodb';
-import memoize from 'fast-memoize';
 import { DynamoDbStore } from './dynamodb-store';
 import { analyticsMetadataTable, prefixMonthGSIName } from '../../common/constants';
 import { docClient, query } from '../../common/dynamodb-client';
@@ -27,7 +26,7 @@ import { ClickStreamStore } from '../click-stream-store';
 import { MetadataStore } from '../metadata-store';
 
 const store: ClickStreamStore = new DynamoDbStore();
-const memoizedQuery = memoize(query);
+const memoizedQuery = query;
 
 export class DynamoDbMetadataStore implements MetadataStore {
 
@@ -90,31 +89,16 @@ export class DynamoDbMetadataStore implements MetadataStore {
 
   public async getEventParameter(projectId: string, appId: string, parameterName: string, valueType: MetadataValueType):
   Promise<IMetadataEventParameter | undefined> {
-    const lastDay = `day${new Date().getDate() - 1}`;
-    const input: QueryCommandInput = {
-      TableName: analyticsMetadataTable,
-      IndexName: prefixMonthGSIName,
-      KeyConditionExpression: '#prefix= :prefix',
-      ProjectionExpression: `#id, #month, #prefix, projectId, appId, #name, eventName, category, valueType, ${lastDay}, summary`,
-      ExpressionAttributeNames: {
-        '#prefix': 'prefix',
-        '#id': 'id',
-        '#month': 'month',
-        '#name': 'name',
-      },
-      ExpressionAttributeValues: {
-        ':prefix': `EVENT_PARAMETER#${projectId}#${appId}`,
-      },
-      ScanIndexForward: false,
-    };
-    let records = await memoizedQuery(input) as IMetadataRaw[];
-    if (records.length === 0) {
-      records = await this.queryEventParameterFromBuiltInList(projectId, appId, parameterName, valueType);
-    }
+    const records = await this.getAllEventParameters(projectId, appId);
     return getParameterByNameAndType(records, parameterName, valueType);
   };
 
   public async listEventParameters(projectId: string, appId: string): Promise<IMetadataEventParameter[]> {
+    const records = await this.getAllEventParameters(projectId, appId);
+    return getLatestParameterById(records);
+  };
+
+  private async getAllEventParameters(projectId: string, appId: string): Promise<IMetadataRaw[]> {
     const lastDay = `day${new Date().getDate() - 1}`;
     const input: QueryCommandInput = {
       TableName: analyticsMetadataTable,
@@ -130,15 +114,14 @@ export class DynamoDbMetadataStore implements MetadataStore {
       ExpressionAttributeValues: {
         ':prefix': `EVENT_PARAMETER#${projectId}#${appId}`,
       },
-      ScanIndexForward: false,
+      ScanIndexForward: true,
     };
     let records = await memoizedQuery(input) as IMetadataRaw[];
     if (records.length === 0) {
       records = await this.queryMetadataRawsFromBuiltInList(projectId, appId, 'EVENT_PARAMETER');
     }
-    const parameters = getLatestParameterById(records);
-    return parameters;
-  };
+    return records;
+  }
 
   public async getUserAttribute(projectId: string, appId: string, userAttributeName: string, valueType: MetadataValueType):
   Promise<IMetadataUserAttribute | undefined> {
@@ -273,78 +256,6 @@ export class DynamoDbMetadataStore implements MetadataStore {
         },
       };
       metadataRaws.push(raw);
-    }
-    return metadataRaws;
-  };
-
-  private async queryEventParameterFromBuiltInList(
-    projectId: string, appId: string, parameterName: string, valueType: MetadataValueType): Promise<IMetadataRaw[]> {
-    const metadataRaws: IMetadataRaw[] = [];
-    const dic = await store.getDictionary('MetadataBuiltInList');
-    if (!dic) {
-      return metadataRaws;
-    }
-    const builtInList = dic.data as IMetadataBuiltInList;
-    const presetEventParameters = builtInList.PresetEventParameters.filter(p => p.name === parameterName && p.dataType === valueType);
-    for (let preset of presetEventParameters) {
-      if (!preset.eventName) {
-        for (let e of builtInList.PresetEvents) {
-          const raw: IMetadataRaw = {
-            id: `${projectId}#${appId}#${e.name}#${preset.category}#${preset.name}#${preset.dataType}`,
-            month: getCurMonthStr(),
-            prefix: `EVENT_PARAMETER#${projectId}#${appId}`,
-            projectId: projectId,
-            appId: appId,
-            name: preset.name,
-            eventName: e.name,
-            category: preset.category,
-            valueType: preset.dataType,
-            summary: {
-              platform: [],
-              hasData: false,
-            },
-          };
-          metadataRaws.push(raw);
-        }
-      } else {
-        const raw: IMetadataRaw = {
-          id: `${projectId}#${appId}#${preset.eventName}#${preset.category}#${preset.name}#${preset.dataType}`,
-          month: getCurMonthStr(),
-          prefix: `EVENT_PARAMETER#${projectId}#${appId}`,
-          projectId: projectId,
-          appId: appId,
-          name: preset.name,
-          eventName: preset.eventName,
-          category: preset.category,
-          valueType: preset.dataType,
-          summary: {
-            platform: [],
-            hasData: false,
-          },
-        };
-        metadataRaws.push(raw);
-      }
-    }
-    const publicEventParameters = builtInList.PublicEventParameters.filter(p => p.name === parameterName && p.dataType === valueType);
-    for (let pub of publicEventParameters) {
-      for (let e of builtInList.PresetEvents) {
-        const raw: IMetadataRaw = {
-          id: `${projectId}#${appId}#${e.name}#${pub.category}#${pub.name}#${pub.dataType}`,
-          month: getCurMonthStr(),
-          prefix: `EVENT_PARAMETER#${projectId}#${appId}`,
-          projectId: projectId,
-          appId: appId,
-          name: pub.name,
-          eventName: e.name,
-          category: pub.category,
-          valueType: pub.dataType,
-          summary: {
-            platform: [],
-            hasData: false,
-          },
-        };
-        metadataRaws.push(raw);
-      }
     }
     return metadataRaws;
   };
